### 2023-01-05 #4919 @simonrp84
Thanks @pdebuyl. This is my first article submitted to JOSS so I'm unfamiliar with the process. Let me know if there's anything needed from my side!
