### 2023-01-13 #5074 @editorialbot
[REVIEW]: pycoxmunk: A python package for computing sea surface reflectance
**Submitting author:** @simonrp84 (<a href="http://orcid.org/0000-0003-3880-6774">Simon Proud</a>)
**Repository:** https://github.com/simonrp84/PyCoxMunk
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.1.0
