### 2023-01-13 #5074 @pdebuyl
Thanks @arthur-e and @molinav for accepting to review! Make sure to issue the command `@editorialbot generate my checklist` (as a comment in this issue, this will trigger the response of @editorialbot).

Feel free to ask questions directly to the author or to me during the review.
