### 2023-01-27 simonrp84/PyCoxMunk#5 @arthur-e
KeyError: "Input scene does not contain reference dataset"
Making progress! I'm not sure if the difficulties I'm experiencing with the `Simple_SLSTR` notebook stem from my lack of experience with `satpy`... But starting off from the changes made after issue #3, I have:

```py
# Create a scene by searching for files
file_list = find_files_and_readers(base_dir = indir, reader = 'slstr_l1b')
scn = Scene(file_list, reader = 'slstr_l1b')

# Load the bands
scn.load(bnames)

# Additionally load the solar and satellite angles
# This saves computing them internally
scn.load(['solar_zenith_angle',
          'solar_azimuth_angle',
          'satellite_zenith_angle',
          'satellite_azimuth_angle'])

# Create the PyCoxMunk class for processing
pcm = PyCoxMunk(scn, bnames)
```

**Unfortunately, I'm getting a `KeyError`:**

```py
KeyError                                  Traceback (most recent call last)
File /usr/local/python-env/PyCoxMunk/lib/python3.8/site-packages/satpy/utils.py:346, in get_satpos(data_arr, preference, use_tle)
    345 try:
--> 346     lon, lat = _get_sat_lonlat(data_arr, lonlat_prefixes)
    347     alt = _get_sat_altitude(data_arr, alt_prefixes)

File /usr/local/python-env/PyCoxMunk/lib/python3.8/site-packages/satpy/utils.py:381, in _get_sat_lonlat(data_arr, key_prefixes)
    380 def _get_sat_lonlat(data_arr, key_prefixes):
--> 381     orb_params = data_arr.attrs["orbital_parameters"]
    382     lon_keys = [prefix + "longitude" for prefix in key_prefixes]

KeyError: 'orbital_parameters'

KeyError                                  Traceback (most recent call last)
File /usr/local/dev/PyCoxMunk2/pycoxmunk/CM_SceneGeom.py:39, in cm_calcangles(inscn, refband)
     38 suna, sunz = _get_sun_angles(inscn[refband])
---> 39 sata, satz = _get_sensor_angles(inscn[refband])
     41 inscn['satellite_azimuth_angle'] = inscn[refband].copy()

File /usr/local/python-env/PyCoxMunk/lib/python3.8/site-packages/satpy/modifiers/angles.py:445, in _get_sensor_angles(data_arr)
    444 preference = satpy.config.get('sensor_angles_position_preference', 'actual')
--> 445 sat_lon, sat_lat, sat_alt = get_satpos(data_arr, preference=preference)
    446 area_def = data_arr.attrs["area"]

File /usr/local/python-env/PyCoxMunk/lib/python3.8/site-packages/satpy/utils.py:354, in get_satpos(data_arr, preference, use_tle)
    353         return _get_satpos_from_platform_name(data_arr)
--> 354     raise KeyError("Unable to determine satellite position. Either the "
    355                    "reader doesn't provide that information or "
    356                    "geolocation datasets were not available.")
    357 return lon, lat, alt

KeyError: "Unable to determine satellite position. Either the reader doesn't provide that information or geolocation datasets were not available."

KeyError                                  Traceback (most recent call last)
Cell In[4], line 2
      1 # Create the PyCoxMunk class for processing
----> 2 pcm = PyCoxMunk(scn, bnames)

File /usr/local/dev/PyCoxMunk2/pycoxmunk/CM_Main.py:132, in PyCoxMunk.__init__(self, scn, band_names, oc_dir, angle_names, do_brdf, mask_bad, delete_when_done)
    130     self.angle_names = angle_names
    131 if angle_names == 'calc':
--> 132     self.scn = cm_calcangles(scn, refband=band_names[0])
    134 if oc_dir is None:
    135     self.use_occci = False

File /usr/local/dev/PyCoxMunk2/pycoxmunk/CM_SceneGeom.py:51, in cm_calcangles(inscn, refband)
     49     inscn['solar_zenith_angle'].data = sunz
     50 except KeyError:
---> 51     raise KeyError("Input scene does not contain reference dataset, please check inputs.")
     52 except ValueError:
     53     raise ValueError("Cannot retrieve solar and satellite angles, please compute manually and specify.")

KeyError: 'Input scene does not contain reference dataset, please check inputs.'
```

**I did check:**

```py
scn.available_dataset_ids()
```
```
[DataID(name='F1', wavelength=WavelengthRange(min=3.55, central=3.74, max=3.93, unit='µm'), resolution=1000, calibration=<calibration.brightness_temperature>, view=<view.nadir>, stripe=<stripe.f>, modifiers=()),
 DataID(name='F1', wavelength=WavelengthRange(min=3.55, central=3.74, max=3.93, unit='µm'), resolution=1000, calibration=<calibration.brightness_temperature>, view=<view.oblique>, stripe=<stripe.f>, modifiers=()),
 DataID(name='F2', wavelength=WavelengthRange(min=10.4, central=10.85, max=11.3, unit='µm'), resolution=1000, calibration=<calibration.brightness_temperature>, view=<view.nadir>, stripe=<stripe.i>, modifiers=()),
 ...
 DataID(name='solar_zenith_angle', resolution=500, view=<view.nadir>, modifiers=()),
 DataID(name='solar_zenith_angle', resolution=500, view=<view.oblique>, modifiers=()),
 DataID(name='solar_zenith_angle', resolution=1000, view=<view.nadir>, modifiers=()),
 DataID(name='solar_zenith_angle', resolution=1000, view=<view.oblique>, modifiers=())]
```

```
