+++ pycoxmunk/Tests/test_SceneGeom.py
@@ -22,6 +22,7 @@ from pycoxmunk.CM_SceneGeom import cm_calcangles
 from unittest import mock
 import numpy as np
 import unittest
+import pytest
 
 
 class TestSceneGeom(unittest.TestCase):
@@ -54,8 +55,9 @@ class TestSceneGeom(unittest.TestCase):
         self.assertAlmostEqual(geom.raa, 20.)
 
         # And likewise for arrays
-        geom = Cm_sg(self.sza, self.saa, self.vza,
-                     self.vaa, self.lats, self.lons)
+        with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
+            geom = Cm_sg(self.sza, self.saa, self.vza,
+                         self.vaa, self.lats, self.lons)
         np.testing.assert_allclose(geom.raa, self.raa)
 
     @mock.patch('satpy.modifiers.angles._get_sensor_angles')
@@ -77,16 +79,16 @@ class TestSceneGeom(unittest.TestCase):
         np.testing.assert_allclose(retr['satellite_azimuth_angle'].data, self.vaa)
 
         # Check what happens if bad refband is passed
-        with self.assertRaises(KeyError):
+        with pytest.raises(KeyError):
             cm_calcangles(test_scn, 'VIS008')
 
         # Check if we pass bad data
-        with self.assertRaises(KeyError):
+        with pytest.raises(KeyError):
             test_scn['VIS006'] = xr.DataArray(np.array([[5., 4.], [4., 5.]]))
             cm_calcangles(test_scn, 'VIS008')
 
         # Check error is raised if angles can't be computed
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             get_sun.side_effect = ValueError("Bad data")
             cm_calcangles(test_scn, 'VIS006')
 
@@ -94,18 +96,21 @@ class TestSceneGeom(unittest.TestCase):
     def test_shapechecker(self):
         """Tests for the shape checker that ensures arrays are same size."""
         # Default case
-        Cm_sg(self.sza, self.saa, self.vza,
-              self.vaa, self.lats, self.lons)
+        with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
+            Cm_sg(self.sza, self.saa, self.vza,
+                  self.vaa, self.lats, self.lons)
 
         # Mismatched arrays
-        with self.assertRaises(AssertionError):
-            Cm_sg(self.sza, self.saa, self.vza,
-                  self.vaa, np.array([1., 2.]), self.lons)
+        with pytest.raises(AssertionError):
+            with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
+                Cm_sg(self.sza, self.saa, self.vza,
+                      self.vaa, np.array([1., 2.]), self.lons)
 
         # One array, rest scalar
         with mock.patch('pycoxmunk.CM_SceneGeom.CMSceneGeom.calc_relazi') as calc_func:
             calc_func.return_value = 1.
-            retr = Cm_sg(self.sza, 5., 3., -10., 15., -21.)
+            with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
+                retr = Cm_sg(self.sza, 5., 3., -10., 15., -21.)
             self.assertEqual(len(retr.sza), 4)
[…]
