+++ pycoxmunk/Tests/test_WaterConsts.py
@@ -18,42 +18,42 @@
 """Test the constants module."""
 
 from pycoxmunk import CM_Constants as CM_C
-import unittest
+import pytest
 
 
-class TestWaterClass(unittest.TestCase):
+class TestWaterClass:
 
     def test_exceptions(self):
         """Check that passing wrong types to the WaterData class raises an exception.
 
         Correct type for the chl_a_coef value is a 2-element list"""
         # Single float
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             CM_C.WaterData(None, None, None, None, None, None, 1.0, None, None)
         # Single item list
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             CM_C.WaterData(None, None, None, None, None, None, [1.0], None, None)
         # Three item list
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             CM_C.WaterData(None, None, None, None, None, None, [1.0, 1.0, 1.0], None, None)
 
     def test_types(self):
         # Check constants are floats."""
-        self.assertTrue(type(CM_C.chl_a_conc) == float)
-        self.assertTrue(type(CM_C.n_air) == float)
+        assert type(CM_C.chl_a_conc) == float
+        assert type(CM_C.n_air) == float
 
     def test_waterdata(self):
         """Check WaterData class works correctly."""
         test_data = CM_C.WaterData(0.5, 1.0, 0.02, 0.1, 0.2, 0.8, [0.1, 0.2], 0.4, 0.6)
-        self.assertTrue(isinstance(test_data, CM_C.WaterData))
-        self.assertTrue(test_data.wavelength == 0.5)
-        self.assertTrue(test_data.total_backscat == 0.6)
-        self.assertTrue(test_data.chl_a_coef[0] == 0.1)
+        assert isinstance(test_data, CM_C.WaterData)
+        assert test_data.wavelength == 0.5
+        assert test_data.total_backscat == 0.6
+        assert test_data.chl_a_coef[0] == 0.1
 
     def test_default_channel_data(self):
         """Check that default water data is correct."""
-        self.assertTrue(CM_C.CM_DATA_DICT[0.47].chl_a_coef[0] == 3.460e-02)
-        self.assertTrue(CM_C.CM_DATA_DICT[1.6].refrac_real == 1.323e+00)
-        self.assertTrue(CM_C.CM_DATA_DICT[3.7].total_abs == 1.223e+04)
-        self.assertTrue(CM_C.CM_DATA_DICT[0.55].wavelength == 0.55)
-        self.assertTrue(CM_C.CM_DATA_DICT[0.87].base_abs == 5.365e+00)
+        assert CM_C.CM_DATA_DICT[0.47].chl_a_coef[0] == 3.460e-02
+        assert CM_C.CM_DATA_DICT[1.6].refrac_real == 1.323e+00
+        assert CM_C.CM_DATA_DICT[3.7].total_abs == 1.223e+04
+        assert CM_C.CM_DATA_DICT[0.55].wavelength == 0.55
+        assert CM_C.CM_DATA_DICT[0.87].base_abs == 5.365e+00
