### 2023-02-06 #5074 @simonrp84
Thanks for the reviews! I've added tests for the `UserWarnings` and have removed the numpy related ones `DeprecationWarnings`. Can't do much about the `numexpr` warning as that comes from an upstream package, but I guess they'll fix it before python 3.12 arrives.

Regarding the examples, the SLSTR example is indeed troublesome - I think something has changed in the satpy library that's causing this but I can't pin it down. Will investigate further and I'll also add an example that uses truly free and open data (probably NOAA's GOES-16) so that you don't need to register to download the example data. I had also thought of hosting the data myself, but that could cause problems in the future so is best avoided, IMO.
