### 2023-02-06 simonrp84/PyCoxMunk#5 @simon-sat
This turned out to be a stupid mistake on my part, if you pull the latest `main` then the SLSTR example should now work!
