### 2023-02-06 2a6c26c
Update license text for clarity.
