+++ pyproject.toml
@@ -12,7 +12,7 @@ version = "1.0.0"
 description = "An algorithm to compute sea surface reflectance using the Cox-Munk approach."
 readme = "README.md"
 authors = [{ name = "Simon R Proud", email = "simon.proud@stfc.ac.uk" }]
-license = { file = "LICENSE" }
+license = { file = "GPLv3" }
 classifiers = [
     "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
     "Programming Language :: Python",
