+++ setup.cfg
@@ -11,7 +11,7 @@ project_urls =
     Bug Tracker = https://github.com/simonrp84/PyCoxMunk/issues
 classifiers =
     Programming Language :: Python :: 3
-    LICENSE :: OSI APPROVED :: GNU GENERAL PUBLIC LICENSE V3 (GPLV3)
+    LICENSE :: OSI Approved :: GNU General Public License v3 (GPLv3)
     Operating System :: OS Independent
 
 [options]
