### 2023-02-06 03f5d61
Fix some warnings in the doc build.
