### 2023-02-06 simonrp84/PyCoxMunk#5 @arthur-e
@simonrp84 I'm making progress! I can run the entire `Example_SLSTR.ipynb` notebook now without error. I see that output `*.tif` files are written to my output directory. This is what the resulting file `cox_munk_refl_S3_20230123_191305.tif` looks like:

![Screenshot from 2023-02-06 07-34-47](https://user-images.githubusercontent.com/1211103/216999712-e3331106-d38f-4ee2-a3eb-b7b45e948312.png)

All of the outputs seem to have data and georeference information except for:

```
S1_20230123_191305.tif
S2_20230123_191305.tif
S3_20230123_191305.tif
```

Each of these is files is too small (ca. 30.8 kb) and has no valid data or georeference information:

```py
rasterio.open('/home/arthur.endsley/Downloads/PyCoxMunk/outputs/S1_20230123_191305.tif')
/usr/local/python-env/PyCoxMunk/lib/python3.8/site-packages/rasterio/__init__.py:304: NotGeoreferencedWarning: Dataset has no geotransform, gcps, or rpcs. The identity matrix will be returned.
  dataset = DatasetReader(path, driver=driver, sharing=sharing, **kwargs)
Out[3]: <open DatasetReader name='/home/arthur.endsley/Downloads/PyCoxMunk/outputs/S1_20230123_191305.tif' mode='r'>
```

If the output reflectance image (above) looks correct, maybe the issue with bands S1, S2, S3 isn't serious(?)
