### 2023-02-06 simonrp84/PyCoxMunk#5 @simon-sat
Hi @arthur-e - can you try updating your copy of the repo? The problem with SLSTR *should* be corrected in the latest commit to the example notebook...
