### 2023-02-06 simonrp84/PyCoxMunk#5 @arthur-e
I cloned a fresh upstream repo and installed again; latest commit was `03f5d618` on Mon Feb 6 14:10:41 2023.

I'll try again after installing with `conda` in a conda environment.
