### 2023-02-06 simonrp84/PyCoxMunk#5 @arthur-e
I'm getting the same result in a clean conda environment, with `pycoxmunk==1.0.0` installed using `conda-forge`. Perhaps there is something wrong with the SLSTR granule I chose? I re-downloaded the granule but am getting the same result. Is there a reference SLSTR dataset I can use?
