### 2023-02-06 #5074 @arthur-e
@simonrp84 Although I still can't get a good result from the SLSTR example, I was able to run the GOES-ABI example and get a nice, specular sun spot image, along with three ABI band images that look good.

I think the last change I would request would be to document the test suite. There are many options for testing in Python, so even though it is very easy to run the suite with `pytest`, it would be good to add this directive to the README and/or the Documentation.
