### 2023-02-08 simonrp84/PyCoxMunk#5 @simon-sat
I think this is because you're installing the `conda-forge` version, which doesn't have the fix pushed to it*. You'll need the current github `main`.

Can you try these steps?
```
conda create --name test_pcm python=3.10
conda activate test_pcm

git clone git@github.com:simonrp84/PyCoxMunk.git
cd PyCoxMunk
pip install -e .
pip install notebook netcdf4 h5netcdf rasterio
```
Then run the notebook. This works in a fresh environment for me, so hopefully it'll work for you too!

\*  I don't want to make a new `PyCoxMunk` release until we know the fixes are working and any other minor problems are corrected. But once we're happy the problem is resolved I'll do a patch release :)
