### 2023-02-10 simonrp84/PyCoxMunk#5 @arthur-e
@simonrp84 With `pip` version 20.0.2 and Python 3.8, it seems it's not possible to install a repo in editable mode unless it has a `setup.py` file.

```sh
conda create --name test_pcm
conda activate test_pcm
pip install -e .
```
```
ERROR: File "setup.py" not found. Directory cannot be installed in editable mode: /usr/local/dev/PyCoxMunk
(A "pyproject.toml" file was found, but editable mode currently requires a setup.py based build.)
```

Independently, I'm having trouble with `conda`, where the `conda` environment's Python binary is not on my path...

```sh
$ conda activate test_pcm
$ which python
/usr/bin/python
```

Not sure why it is giving me the system-wide Python installation. I get a similar result with `which python3`. I don't have this problem with `virtualenv`:

```sh
$ git clone git@github.com:simonrp84/PyCoxMunk.git
$ cd PyCoxMunk
$ virtualenv /usr/local/python-env/PyCoxMunk/
$ source /usr/local/python-env/PyCoxMunk/bin/activate
$ which python
/usr/local/python-env/PyCoxMunk/bin/python
```

**Anyways, after installing a `virtualenv` with `pip` I'm getting the same results;** everything in the `Simple_SLSTR.ipynb` notebook runs smoothly, without error, but the S1, S2, S3 band GeoTIFFs are each only about ~30 kb and seem to contain no valid data. The Cox-Munk reflectance files are each a smooth, horizontal gradient.

```sh
cd /usr/local/dev/PyCoxMunk
pip install -e .
pip install notebook netcdf4 h5netcdf rasterio
jupyter notebook
```

This is with the PyCoxMunk trunk and:

```
numpy==1.23.5
scipy==1.10.0
pandas==1.5.3
dask==2023.2.0
pyorbital==1.7.3
pyresample==1.26.1
satpy==0.40.0
click==8.1.3
numba==0.56.4
xarray==2023.1.0
```

I don't want to hold up [the JOSS review](https://github.com/openjournals/joss-reviews/issues/5074#issuecomment-1419293131). I doubt that it matters whether `conda` or `virtualenv` are used but...who knows? If this Notebook is working for @molinav then it must just be a problem with my setup. Thank goodness, for my own sake, I only do terrestrial remote sensing. :sweat_smile:

EDIT: In general, I try not to mix `pip` and `conda`, but I understand that there is not another way to install a local package in a `conda` environment(?) If you add a `setup.py` file, I could use `python setup.py install` within the repo. As you already have a `setup.cfg` file, the `setup.py` could be quite minimal, [as in this example](https://github.com/arthur-e/pyl4c/blob/master/setup.py). In fact, I guess I can try adding it myself...
