### 2023-02-10 simonrp84/PyCoxMunk#5 @simon-sat
Thanks for the info. I'll be away for the next few days but will investigate more towards the end of next week.

Can you confirm that the GOES example works? You should be able to test that with the `PyCoxMunk` from conda / pip (without needing editable mode) and just downloading the example notebook itself from github...
