### 2023-02-10 simonrp84/PyCoxMunk#5 @arthur-e
@simonrp84 Yes, the GOES-ABI example works great for me. I did not get through `Complex_SEVIRI.ipynb` because I didn't have ERA5 re-analysis data handy.

I did figure out how to get Python into my `conda` environment. I had avoided the `python=` argument because I'm using a different version of Python; I figured `conda` would create my environment with the default Python installation (as is the case with `virtualenv`) but, in fact, it seems it must be expressly requested:

```sh
$ conda create -n PyCoxMunk python=python3.8
$ conda activate PyCoxMunk
$ which python
/usr/local/dev/miniconda3/envs/PyCoxMunk/bin/python
$ which pip
/usr/local/dev/miniconda3/envs/PyCoxMunk/bin/pip
```

Success with installing again, though I needed to activate Jupyter Notebook differently...

```sh
pip install -e .
pip install notebook netcdf4 h5netcdf rasterio
python -m notebook
```

Yet the result for `Example_SLSTR.ipynb` is the same, unfortunately.

I'm happy to move the review forward if you are able to add documentation on the test suite, when you return.
