### 2023-02-21 #5074 @molinav
Hi, @simonrp84! I re-checked everything after re-building the library from the latest `main` branch. I come with a second round of comments:

- About the manuscript: just two minor text edits.

  - [ ] Line 78: no comma required between "main class" and "PyCoxMunk" (or surround "PyCoxMunk" with two commas).
  - [ ] Line 80: heigh => height (or high? I am not a native speaker).

- About the GitHub workflow:

  - For a pure Python library like `pycoxmunk` it is not something dramatic, but it is good practice to run the test suite against the *installed* library instead of the cloned repository folder. This way, you ensure that the generated sdists and wheels are consistent and not broken. From former experiences (making this mistake myself), your library could depend on auxiliary data files that are not packed into the sdist/wheel properly, and your tests would pass because this data is in the cloned repo, but the distributable package would be broken because of not having this data. The workflow update should contain something like below before `pytest` (if using the new `build` mechanism, as there is no legacy setup file):
```sh
python -m pip install build
python -m build
python -m pip install dist/*.whl
```

- About the docs:

  - Just for reference, continuing about the automation of docs in the future, I can recommend that you look at how big scientific libraries are doing, the docstring syntax is easy to follow (see e.g. the [`numpy.take`](https://github.com/numpy/numpy/blob/v1.24.2/numpy/core/fromnumeric.py#L94) function docstring).

- About the license:

  - [ ] The license metadata needs one additional correction. It is still saying `license = { file = "GPLv3" }`, but after the first correction this attribute does not refer to a file anymore. Therefore it should say `license = { text = "GPLv3" }` (according to the [example](https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html) in the `setuptools` docs).
  - [ ] From what I see in the top-level comment of each module, it looks like you would allow to use the library also with any version after GPLv3. However, I think that the GPLv3 classifier refers to the GPL-3.0-only variant (note that there is a `License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)` classifier, see [here](https://autopilot-docs.readthedocs.io/en/latest/license_list.html)). Due to the special nature of the GPL-like licenses, I would be cautious about setting this attribute consistently everywhere. I had some issues in the past, e.g. trying to resolve the license implications about GEOS because the license and library headers say one thing (ambiguity if it is LGPLv2.1 or LGPLv2.1+), but the main author says a different thing (LGPLv2.1 only).

- About the tests:

  - [ ] The `pyproject.toml` now has an update to ignore `RuntimeWarning` when running `pytest`. Using `"ignore::RuntimeWarning"` is too broad and may hide you unexpected warnings in the future. If you are trying to skip the warning coming from `da.arccos`, I would replace the filter to `"ignore::RuntimeWarning:pycoxmunk.CM_Shared_Wind:"` to narrow down a bit. Still, I would check if this `RuntimeWarning` is coming from passing something to `da.arccos` outside [-1, 1], because in that case the warning is expected and should be caught as part of the running test suite. Or, if you want to gracefully return nans without warnings under these circumstances, the warning filtering should be performed inside `CMSharedWind.calc_wind`.
  - [ ] I need to check it deeper, but on Windows I am getting errors when running `pytest` due to warnings not being thrown. Tested with Python 3.9 and Python 3.10, see the tracebacks below. As said, I will review this a bit more on my side, because your GitHub workflow is running on Ubuntu with Python 3.10 and the tests are passing, so it might be an OS-specific issue or I might have done something wrong when preparing the virtual environments. The tracebacks below correspond to the current main branch, if I move to the v1.0.0 tag then the tests are passing (with the warnings mentioned in my first review), so something in between may be causing this change.
```
(joss39) [vic@surface] C:\Users\vic\Desktop\scratch\PyCoxMunk> python -m pytest --cov=pycoxmunk --cov-report html pycoxmunk\Tests
================================================= test session starts ==================================================
platform win32 -- Python 3.9.13, pytest-7.2.1, pluggy-1.0.0
rootdir: C:\Users\vic\Desktop\scratch\PyCoxMunk, configfile: pyproject.toml
plugins: cov-4.0.0

pycoxmunk\Tests\test_CMCalcs.py .........                                                                         [ 28%]
pycoxmunk\Tests\test_CM_main.py .....FFFF                                                                         [ 56%]
pycoxmunk\Tests\test_PixMask.py ..                                                                                [ 62%]
pycoxmunk\Tests\test_SceneGeom.py ....                                                                            [ 75%]
pycoxmunk\Tests\test_SharedWind.py .                                                                              [ 78%]
pycoxmunk\Tests\test_Utils.py ...                                                                                 [ 87%]
pycoxmunk\Tests\test_WaterConsts.py ....                                                                          [100%]

======================================================= FAILURES =======================================================
______________________________________________ TestCMMain.test_wind_setup ______________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000179D68EF8E0>

    def test_wind_setup(self):
        """Test that winds are set up correctly."""
        mocker = mock.MagicMock()
        mocker.return_value = True
        with mock.patch('pycoxmunk.CM_Main.CMSharedWind', mocker):
            with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
>               pcm = PyCoxMunk(self.scn_good, self.good_bnd_names)
E               Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E               The list of emitted warnings is: [].

pycoxmunk\Tests\test_CM_main.py:168: Failed
_______________________________________________ TestCMMain.test_pixmask ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000179D68EFA30>

    def test_pixmask(self):
        """Test that pixmask is correctly initialised in the class."""
        with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
>           pcm = PyCoxMunk(self.scn_good, self.good_bnd_names)
E           Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E           The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk\Tests\test_CM_main.py:193: Failed
_______________________________________________ TestCMMain.test_retr_cm ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000179D68EF9D0>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='1622854006144'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_retr_cm(self, mock_cmr_func):
        """Tests for the main routine that retrieves reflectance."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        # This section tests that cox_munk reflectance is computed, but not BRDF
        # Recreate scene in case previous tests have messed with it.
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
        with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
>           pcm = PyCoxMunk(self.scn_good, self.good_bnd_names, mask_bad=False, do_brdf=False)
E           Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E           The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk\Tests\test_CM_main.py:210: Failed
_______________________________________________ TestCMMain.test_deleter ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000179D68EFB20>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='1622854492224'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_deleter(self, mock_cmr_func):
        """Test that deletions are done correctly."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        tmp_arr = np.array([1, 4., 64])

---------- coverage: platform win32, python 3.9.13-final-0 -----------
Coverage HTML written to dir htmlcov

=============================================== short test summary info ================================================
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_wind_setup - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_pixmask - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_retr_cm - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_deleter - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
```
```
(joss310) [vic@surface] C:\Users\vic\Desktop\scratch\PyCoxMunk> python -m pytest --cov=pycoxmunk --cov-report html pycoxmunk\Tests
================================================= test session starts ==================================================
platform win32 -- Python 3.10.10, pytest-7.2.1, pluggy-1.0.0
rootdir: C:\Users\vic\Desktop\scratch\PyCoxMunk, configfile: pyproject.toml
plugins: cov-4.0.0

pycoxmunk\Tests\test_CMCalcs.py .........                                                                         [ 28%]
pycoxmunk\Tests\test_CM_main.py .....FFFF                                                                         [ 56%]
pycoxmunk\Tests\test_PixMask.py ..                                                                                [ 62%]
pycoxmunk\Tests\test_SceneGeom.py ....                                                                            [ 75%]
pycoxmunk\Tests\test_SharedWind.py .                                                                              [ 78%]
pycoxmunk\Tests\test_Utils.py ...                                                                                 [ 87%]
pycoxmunk\Tests\test_WaterConsts.py ....                                                                          [100%]

======================================================= FAILURES =======================================================
______________________________________________ TestCMMain.test_wind_setup ______________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x000001CDF7EBE200>

    def test_wind_setup(self):
        """Test that winds are set up correctly."""
        mocker = mock.MagicMock()
        mocker.return_value = True
        with mock.patch('pycoxmunk.CM_Main.CMSharedWind', mocker):
>           with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E           Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E           The list of emitted warnings is: [].

pycoxmunk\Tests\test_CM_main.py:167: Failed
_______________________________________________ TestCMMain.test_pixmask ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x000001CDF7EBDDE0>

    def test_pixmask(self):
        """Test that pixmask is correctly initialised in the class."""
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk\Tests\test_CM_main.py:192: Failed
_______________________________________________ TestCMMain.test_retr_cm ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x000001CDF7EBDEA0>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='1984191447824'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_retr_cm(self, mock_cmr_func):
        """Tests for the main routine that retrieves reflectance."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        # This section tests that cox_munk reflectance is computed, but not BRDF
        # Recreate scene in case previous tests have messed with it.
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk\Tests\test_CM_main.py:209: Failed
_______________________________________________ TestCMMain.test_deleter ________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x000001CDF7EBDB40>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='1984158534480'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_deleter(self, mock_cmr_func):
        """Test that deletions are done correctly."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        tmp_arr = np.array([1, 4., 64])

        # Test for case where we don't delete intermediate data
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk\Tests\test_CM_main.py:268: Failed

---------- coverage: platform win32, python 3.10.10-final-0 ----------
Coverage HTML written to dir htmlcov

=============================================== short test summary info ================================================
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_wind_setup - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_pixmask - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_retr_cm - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_deleter - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
============================================ 4 failed, 28 passed in 50.05s =============================================
```

- About `Simple_GOES_ABI.ipynb`: thanks for adding this example, I was able to run it without issues. I have two suggestions:

  - [ ] "An library is required for this notebook [...]" => "The following libraries are required for this notebook: `h5netcdf`, `rasterio` and `s3fs`". The first two libraries were not mentioned and only realised when getting `ImportError` at runtime.
  - [ ] "We do not apply and cloud or land masking" => "We do not apply any cloud or land masking".

- About `Simple_SLSTR.ipynb`:

  - [ ] For me, this Jupyter notebook still does not work. The error occurs at the same cell as in the first round. I also realised that this notebook has a hidden dependency. Is it expected to have `netCDF4`, `h5netcdf`, or any of them?
# Create the PyCoxMunk class for processing
pcm = PyCoxMunk(scn, bnames)
KeyError                                  Traceback (most recent call last)
File c:\Users\vic\Desktop\scratch\joss310\lib\site-packages\satpy\utils.py:345, in get_satpos(data_arr, preference, use_tle)
--> 345     lon, lat = _get_sat_lonlat(data_arr, lonlat_prefixes)
    346     alt = _get_sat_altitude(data_arr, alt_prefixes)

File c:\Users\vic\Desktop\scratch\joss310\lib\site-packages\satpy\utils.py:380, in _get_sat_lonlat(data_arr, key_prefixes)
    379 def _get_sat_lonlat(data_arr, key_prefixes):
--> 380     orb_params = data_arr.attrs["orbital_parameters"]
    381     lon_keys = [prefix + "longitude" for prefix in key_prefixes]

KeyError: 'orbital_parameters'

KeyError                                  Traceback (most recent call last)
File ~\Desktop\scratch\PyCoxMunk\pycoxmunk\CM_SceneGeom.py:39, in cm_calcangles(inscn, refband)
     38 suna, sunz = _get_sun_angles(inscn[refband])
---> 39 sata, satz = _get_sensor_angles(inscn[refband])
     41 inscn['satellite_azimuth_angle'] = inscn[refband].copy()

File c:\Users\vic\Desktop\scratch\joss310\lib\site-packages\satpy\modifiers\angles.py:447, in _get_sensor_angles(data_arr)
    446 preference = satpy.config.get('sensor_angles_position_preference', 'actual')
--> 447 sat_lon, sat_lat, sat_alt = get_satpos(data_arr, preference=preference)
...
---> 51     raise KeyError("Input scene does not contain reference dataset, please check inputs.")
     52 except ValueError:
     53     raise ValueError("Cannot retrieve solar and satellite angles, please compute manually and specify.")

KeyError: 'Input scene does not contain reference dataset, please check inputs.'
```
