### 2023-02-21 #5074 @pdebuyl
Hi all, sorry for not checking on the review .Thanks @arthur-e and @molinav for the review so far!

@simonrp84 I also had trouble with the SLSTR example. Even though the data requires registration, could you provide a "ready to type" query to fetch the data? We don't have those via EUMETCast at my institution, so I used https://scihub.copernicus.eu and picked a region at random and SLSTR as the product. I'll check again when at work though since you update the code.
