### 2023-03-21 simonrp84/PyCoxMunk#5 @simon-sat
Hi,  I've finally come back to look at PyCoxMunk after some time on other projects. Have just tested on a completely fresh virtual machine:

```
wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh
bash Miniconda3-latest-Linux-x86_64.sh
[ install conda]
conda install mamba
mamba install pycoxmunk jupyter
```
I then downloaded the `Example_SLSTR.ipynb` notebook and a sample SLSTR file from https://dataspace.copernicus.eu/
Then I ran the notebook, and the results were as expected. So unfortunately I cannot reproduce the problem you have experienced here and am not sure what the best approach to resolve it is.
But I am happy that the GOES notebook at least worked successfully!
