### 2023-03-21 #5074 @simonrp84
@pdebuyl I have now added another example, using GOES/ABI data, that should work out-of-the-box without the user having to download any files. It's all done within the notebook itself.

After some time having to work on other things I'm now able to look at PyCoxMunk again and hope to address all the review comments in the next day or two.
