### 2023-03-21 93fe25f
Update readme to provide some basic info on the tests.
