### 2023-03-21 #5074 @simonrp84
@arthur-e Thanks for the suggestion regarding some documentation for the tests. I've now added a brief section to the readme and a page to readthedocs  that describes how to run / use the tests.
