+++ docs/requirements.txt
@@ -1,5 +1,4 @@
 # File: docs/requirements.txt
 
 sphinx>=4.2.0
-sphinx_rtd_theme>=1.0.0
 readthedocs-sphinx-search==0.1.1
