+++ pycoxmunk/Tests/test_SharedWind.py
@@ -60,4 +60,3 @@ class TestSharedWind:
         np.testing.assert_allclose(sh_wind.cosbeta, self.cosbeta, rtol=1e-5)
         np.testing.assert_allclose(sh_wind.w, self.w, rtol=1e-5)
         np.testing.assert_allclose(sh_wind.p, self.p, rtol=1e-5)
-
