### 2023-03-21 cd6b79e
Update license text to specify GPL v3 only.
