### 2023-03-21 #5074 @simonrp84
@molinav Thanks for your additional comments.

 - paper:
I hadn't spotted the two typos you mention in the paper - have corrected those now!

- workflow:
Interesting, I've added that to the workflow file and tests are passing, so hopefully that's been a simple change that could help in the future.

- docs:
Thanks for the suggestion. I'll definitely bear that in mind for future projects; seems like a timesaver.

- license:
Yes, you're right. I've now corrected those inconsistencies across the various files to that v3 only is specifies and have also changed `file` to `text`.

- RuntimeWarning :
I have removed this code from the `pyproject.toml` again - I think I added that for testing and forgot to remove it. The tests aren't producing a RuntimeWarning now.

- Windows test errors:
I can't reproduce this, unfortunately. I did all the development on windows and didn't notice any errors like this, and have also just tried to reproduce on a fresh windows VM with no success, the tests run OK.
Can you confirm if you still see these errors?

- Dependencies for `Simple_GOES_ABI.ipynb`:
You're right, I thought these were dependencies of satpy and hence installed as pycoxmunk requires satpy. But they're optional rather than required dependencies in satpy. Have now added these to the notebook.

- Problems with `Simple_SLSTR.ipynb`:
Just to check, are you trying to run the latest version from github `main` branch or the one from the conda / pypi release? I've not made a new release yet so the latter ones will still be broken. For me, the notebook on github works OK on a clean conda environment for both windows and linux.
(steps I took are listed here: https://github.com/simonrp84/PyCoxMunk/issues/5 )

As long as you have the notebook from github, there's no need to install pycoxmunk from github. The version on conda works fine.
Please let me know if you still experience problems with the notebook and I'll dig deeper.
