### 2023-04-04 #5074 @pdebuyl
For info if someone else is interested:

```
import cdsapi

c = cdsapi.Client()

c.retrieve(
    'reanalysis-era5-single-levels',
    {
        'product_type': 'reanalysis',
        'variable': [
            '10m_u_component_of_wind', '10m_v_component_of_wind',
        ],
        'year': '2016',
        'month': '03',
        'day': '26',
        'time': '15:00',
        'format': 'netcdf',
    },
    'download.nc')
```

will provide the ERA5 netcdf data. (requires an account with the cds and `pip install cdsapi`)
