### 2023-04-04 #5074 @pdebuyl
```
================================================================================ test session starts =================================================================================
rootdir: /tmp/PyCoxMunk
plugins: anyio-3.6.2
collected 32 items

pycoxmunk/Tests/test_CMCalcs.py .........                                                                                                                                      [ 28%]
pycoxmunk/Tests/test_CM_main.py .....FFFF                                                                                                                                      [ 56%]
pycoxmunk/Tests/test_PixMask.py ..                                                                                                                                             [ 62%]
pycoxmunk/Tests/test_SceneGeom.py ....                                                                                                                                         [ 75%]
pycoxmunk/Tests/test_SharedWind.py .                                                                                                                                           [ 78%]
pycoxmunk/Tests/test_Utils.py ...                                                                                                                                              [ 87%]
pycoxmunk/Tests/test_WaterConsts.py ....                                                                                                                                       [100%]

====================================================================================== FAILURES ======================================================================================
_____________________________________________________________________________ TestCMMain.test_wind_setup _____________________________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x7f30cba85c90>

    def test_wind_setup(self):
        """Test that winds are set up correctly."""
        mocker = mock.MagicMock()
        mocker.return_value = True
        with mock.patch('pycoxmunk.CM_Main.CMSharedWind', mocker):
>           with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E           Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E           The list of emitted warnings is: [].

pycoxmunk/Tests/test_CM_main.py:165: Failed
______________________________________________________________________________ TestCMMain.test_pixmask _______________________________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x7f30cba85b10>

    def test_pixmask(self):
        """Test that pixmask is correctly initialised in the class."""
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk/Tests/test_CM_main.py:190: Failed
______________________________________________________________________________ TestCMMain.test_retr_cm _______________________________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x7f30cba85810>, mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='139847227143648'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_retr_cm(self, mock_cmr_func):
        """Tests for the main routine that retrieves reflectance."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        # This section tests that cox_munk reflectance is computed, but not BRDF
        # Recreate scene in case previous tests have messed with it.
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk/Tests/test_CM_main.py:207: Failed
______________________________________________________________________________ TestCMMain.test_deleter _______________________________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x7f30cba84940>, mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='139847226750240'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_deleter(self, mock_cmr_func):
        """Test that deletions are done correctly."""

        tmp_arr = np.array([1, 4., 64])

        # Test for case where we don't delete intermediate data
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in double_scalars')].

pycoxmunk/Tests/test_CM_main.py:262: Failed
================================================================================== warnings summary ==================================================================================
pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_badbands
pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_ocean_color_init
pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_ocean_color_init
pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_calcangles
  /tmp/PyCoxMunk/pycoxmunk/CM_Shared_Wind.py:68: RuntimeWarning: invalid value encountered in double_scalars
    self.wd = da.arccos(self.v10 / self.ws)

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================================================== short test summary info ===============================================================================
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_wind_setup - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_pixmask - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_retr_cm - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_deleter - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
===================================================================== 4 failed, 28 passed, 4 warnings in 12.37s ======================================================================
```

I also have the failing tests (linux Python 3.10)
