### 2023-04-04 #5074 @pdebuyl
@arthur-e @molinav would you assess the latest update by @simonrp84 ?

(PS: @simonrp84 would you provide the example output for the example notebooks?)
