### 2023-05-01 #5074 @arthur-e
I can verify that all tests pass on the trunk version (1d57ab). However, I found it was necessary to use:

```py
pytest-3 .
```

As `pytest` seems to be reserved for the Python 2 version of `pytest`. Maybe this depends on how `pytest` is installed. I didn't use `conda`.

All of my review items are complete and I can recommend `PyCoxMunk` be accepted.
