### 2023-05-02 #5074 @simonrp84
@arthur-e Thanks for checking. I suspect that may be something to do with the way it's installed. `pytest .` works OK here (win32, all packages installed via conda).
