### 2023-05-02 #5074 @molinav
@simonrp84 I will try to provide my final feedback between today and tomorrow, I just came back from holidays. I will have no objections with having `PyCoxMunk` being accepted, most of my last points were minor details and I can always report them to you in the `PyCoxMunk` issue tracker.
