### 2023-05-02 #5074 @simonrp84
Just to add, I plan some minor updates to the example code (fixing a few incorrect comments) and I'll also add the ERA5 data to the repository.

You asked if I could upload the example notebook output, @pdebuyl: I'm sorting that now and plan to put it on zenodo. Will see if I can include the SEVIRI data file used in the notebook too, but need to hear back from EUMETSAT about the license first.
