### 2023-05-02 #5074 @simonrp84
OK, @pdebuyl I have added sample output data for the three example notebooks to a zenodo repository, and have also included an SLSTR scene in the same repo: https://zenodo.org/record/7885996

I would also like to include SEVIRI data in that, but need to check with EUMETSAT about licensing. Not sure how long that'll take, hopefully it's quick! If not, I can always add SEVIRI to the repo once I have a response from them in the future.
