### 2023-05-02 #5074 @simonrp84
Based on feedback from EUMETSAT, I have now been able to include a SEVIRI dataset on zenodo for use with the example notebook. Have updated the notebooks accordingly.

I also discovered why you have noticed tests failing. I had some tests set up to calculate angles internally, and floating point errors were meaning tests occasionally failed (happened on my laptop, but not my office PC). This should also be corrected in the latest push to `main`.
