### 2023-05-04 #5074 @molinav
I re-checked everything with the latest `main` branch of `pycoxmunk`.

- The requests in the manuscript have been addressed.

- The requests in the GitHub workflow have also been addressed (see [GitHub Workflow]). I think the workflow change to `python -m build` is not enough here to test against the installed wheel, because the tests are run in the repository root folder and the `pycoxmunk` folder is in `PATH` and might take precedence over the installed wheel. Anyway, it is just a detail and I can give you advice outside the paper loop.

[GitHub Workflow]:
https://github.com/simonrp84/PyCoxMunk/blob/a0ee002/.github/workflows/python-app.yaml#L38

- The requests about the license have been clarified successfully.

- About the unit tests, I keep getting errors on Windows (this time with Python 3.11 and `pyenv-win` Python installation, no `conda` involved). All errors are about expected warnings not being found. This is similar to what I found in previous iterations. These errors in the unit tests are not as dramatic as if they were caused by unexpected results. Still I would consider important to fix them at some point, because they might be source of corner situations that later might be difficult to debug. If @pdebuyl is ok with it, we can move forward here and we continue iterating in your issue tracker, since I do not know the criterion of JOSS in this situation. I leave the traceback below for reference:

```
(joss311) [vic@surface] C:\Users\vic\Desktop\scratch> python -m pytest --cov=pycoxmunk --cov-report html PyCoxMunk/pycoxmunk/Tests
========================================================== test session starts ===========================================================
platform win32 -- Python 3.11.2, pytest-7.3.1, pluggy-1.0.0
rootdir: C:\Users\vic\Desktop\scratch
plugins: cov-4.0.0
collected 32 items

PyCoxMunk\pycoxmunk\Tests\test_CMCalcs.py .........                                                                                 [ 28%]
PyCoxMunk\pycoxmunk\Tests\test_CM_main.py ...F.FFFF                                                                                 [ 56%]
PyCoxMunk\pycoxmunk\Tests\test_PixMask.py ..                                                                                        [ 62%]
PyCoxMunk\pycoxmunk\Tests\test_SceneGeom.py ....                                                                                    [ 75%]
PyCoxMunk\pycoxmunk\Tests\test_SharedWind.py .                                                                                      [ 78%]
PyCoxMunk\pycoxmunk\Tests\test_Utils.py ...                                                                                         [ 87%]
PyCoxMunk\pycoxmunk\Tests\test_WaterConsts.py ....                                                                                  [100%]

================================================================ FAILURES ================================================================
_______________________________________________________ TestCMMain.test_calcangles _______________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000236D1AB1350>

    def test_calcangles(self):
        """Check that angle calculation is initialised correctly."""
        mocker = mock.MagicMock()
        mocker_lalo = mock.MagicMock()
        mocker_lalo.attrs['area'].get_lonlats = self._get_lonlats
        tmp_dict = {'VIS006': mocker_lalo,
                    'solar_zenith_angle': np.array([10, 10]),
                    'satellite_zenith_angle': np.array([10, 10]),
                    'solar_azimuth_angle': np.array([10, 10]),
                    'satellite_azimuth_angle': np.array([10, 10])}
        mocker.return_value = tmp_dict
        with mock.patch('pycoxmunk.CM_Main.cm_calcangles', mocker):
>           with pytest.warns(RuntimeWarning, match="invalid value encountered in double_scalars"):
E           Failed: DID NOT WARN. No warnings of type (<class 'RuntimeWarning'>,) matching the regex were emitted.
E            Regex: invalid value encountered in double_scalars
E            Emitted warnings: [RuntimeWarning('invalid value encountered in scalar divide')]

PyCoxMunk\pycoxmunk\Tests\test_CM_main.py:170: Failed
_______________________________________________________ TestCMMain.test_wind_setup _______________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000236D1AB2310>

    def test_wind_setup(self):
        """Test that winds are set up correctly."""
        mocker = mock.MagicMock()
        mocker.return_value = True
        with mock.patch('pycoxmunk.CM_Main.CMSharedWind', mocker):
>           with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E           Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E           The list of emitted warnings is: [].

PyCoxMunk\pycoxmunk\Tests\test_CM_main.py:185: Failed
________________________________________________________ TestCMMain.test_pixmask _________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000236D1AB0C90>

    def test_pixmask(self):
        """Test that pixmask is correctly initialised in the class."""
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in scalar divide')].

PyCoxMunk\pycoxmunk\Tests\test_CM_main.py:210: Failed
________________________________________________________ TestCMMain.test_retr_cm _________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000236D1AB2FD0>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='2434511232848'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_retr_cm(self, mock_cmr_func):
        """Tests for the main routine that retrieves reflectance."""

        cm_band_name = f'cox_munk_refl_{self.good_bnd_names[0]}'
        cm_rho0v_name = f'cox_munk_rho0v_{self.good_bnd_names[0]}'

        # This section tests that cox_munk reflectance is computed, but not BRDF
        # Recreate scene in case previous tests have messed with it.
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in scalar divide')].

PyCoxMunk\pycoxmunk\Tests\test_CM_main.py:227: Failed
________________________________________________________ TestCMMain.test_deleter _________________________________________________________

self = <pycoxmunk.Tests.test_CM_main.TestCMMain object at 0x00000236D1AB35D0>
mock_cmr_func = <MagicMock name='calc_coxmunk_wrapper' id='2434506064080'>

    @mock.patch('pycoxmunk.CM_Main.calc_coxmunk_wrapper')
    def test_deleter(self, mock_cmr_func):
        """Test that deletions are done correctly."""

        tmp_arr = np.array([1, 4., 64])

        # Test for case where we don't delete intermediate data
        self.scn_good = self.create_test_scene(self.good_bnd_names, self.good_angle_names)
        mock_cmr_func.return_value = self._make_mocked_cm_refl()
>       with pytest.warns(UserWarning, match="Some solar zenith values out of range. Clipping."):
E       Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
E       The list of emitted warnings is: [RuntimeWarning('invalid value encountered in scalar divide')].

PyCoxMunk\pycoxmunk\Tests\test_CM_main.py:282: Failed

---------- coverage: platform win32, python 3.11.2-final-0 -----------
Coverage HTML written to dir htmlcov
======================================================== short test summary info =========================================================
FAILED PyCoxMunk/pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_calcangles - Failed: DID NOT WARN. No warnings of type (<class 'RuntimeWarning'>,) matching the regex were emitted.
FAILED PyCoxMunk/pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_wind_setup - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED PyCoxMunk/pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_pixmask - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED PyCoxMunk/pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_retr_cm - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
FAILED PyCoxMunk/pycoxmunk/Tests/test_CM_main.py::TestCMMain::test_deleter - Failed: DID NOT WARN. No warnings of type (<class 'UserWarning'>,) were emitted.
===================================================== 5 failed, 27 passed in 25.25s ======================================================
```

- About `Simple_GOES_ABI.ipynb`: working as before. One typo was not corrected from the previous iteration:

  - [ ] "We do not apply and cloud or land masking" => "We do not apply any cloud or land masking".

- About `Simple_SLSTR.ipynb`: now working. Here I used again the file that I downloaded, "S3A_SL_1_RBT____20230203T041025_20230203T041325_20230204T123946_0179_095_104_3060_PS1_O_NT_004.SEN3". Same typo correction applies as before:

  - [ ] "We do not apply and cloud or land masking" => "We do not apply any cloud or land masking".

- About `Complex_SEVIRI.ipynb`:

  - `netCDF4` needs to be indicated as additional library to install for this example, since it is not installed as a dependency of `pycoxmunk` and the user might not have it.

  - It seems that `xarray==2023.3.0` is the last version of `xarray` that can identify `rasterio` as a backend to open generic image files. From version `xarray==2023.4.0`, this is not the case anymore (see [`xarray` release notes for v2023.4.0](https://github.com/pydata/xarray/releases/tag/v2023.04.0)) and the IPython notebook will fail when trying to open the SEVIRI land mask in TIFF format. It seems that this backend was deprecated for some time and it just happened that they removed this functionality during the review loop of your paper. I do not have a solution for that, as I am not an `xarray` regular user. If I test the notebook skipping the land mask loading and the `pcm.setup_pixmask` call, I can reach the end of the notebook and save the output files, but the results differ of course.

TLDR summary comes in a following comment.
