### 2023-05-04 #5074 @molinav
So my overall summary would be the following:
- The purpose of the library is clear, the base functionality is clear, the licensing is clear, the packaging is correct.
- The GitHub workflow would need minor tweaking to ensure that the tests are run for the installed wheel.
- The GitHub workflow could add in the future a matrix of Python versions and operating systems to avoid having unexpected behaviours (e.g. why sometimes the unit tests do not throw the expected warnings).
- The three example notebooks essentially work. They would need a bit of improvement so that they clearly state the extra dependencies that are required. Also, the notebooks define specific paths based on the "D:" Windows drive, which force the user to tune the notebooks before they can actually work (e.g. these paths are not valid for a GNU/Linux user). Other issues are coming from changes in external libraries, and this is out of control of `pycoxmunk`.

If I had to decide, I would say the `pycoxmunk` paper can be accepted now as it is. All these corrections coming from me are in the end minor corrections, and if `pycoxmunk` keeps alive they would get fixed sooner or later. No 1.0.0 library version is perfect anyway, and @simonrp84 has been actively updating the library based on the feedback received here, so I am sure that the remaining issues will eventually be solved in the corresponding GitHub repo.
