### 2023-05-04 #5074 @simonrp84
Thanks @molinav! Not sure how I overlooked those typos in the notebooks. Fixed now. I also added a note to the SEVIRI notebook saying that `netCDF4` is required. I also added a note to all three notebooks discussing the directories, and changed the default directory from `D:/sat_data/...` to `./`, which should be cross-platform.

Regarding your xarray comment: This is an issue in satpy, we released a new version last week (0.42.x) that should fix it, so I guess you're using a slightly older version of satpy. I just tried with a fresh environment and it works OK with the latest versions.

The unit tests issue is interesting - I've been testing on python <3.10 and actually installed 3.11 for the first time today. Running the tests now I can reproduce your failed warnings so presumably something has changed in the new python version. I'll investigate that...and agree it'd be good to test in CI against multiple versions.
