### 2023-05-04 5ffc81f
Update tests to ensure warnings are raised when they should be.
