### 2023-05-05 #5074 @molinav
Hi @simonrp84! I can also confirm that the tests are passing for me after your last changes. See traceback below:
```
(joss311) [vic@onyx] C:\Users\vic\Desktop\scratch> python -m pytest --cov=pycoxmunk --cov-report=html .\libraries\PyCoxMunk\pycoxmunk\Tests
================================================================================= test session starts ==================================================================================
platform win32 -- Python 3.11.2, pytest-7.3.1, pluggy-1.0.0
rootdir: C:\Users\vic\Desktop\scratch
plugins: cov-4.0.0
collected 32 items

libraries\PyCoxMunk\pycoxmunk\Tests\test_CMCalcs.py .........                                                                                                                     [ 28%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_CM_main.py .........                                                                                                                     [ 56%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_PixMask.py ..                                                                                                                            [ 62%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_SceneGeom.py ....                                                                                                                        [ 75%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_SharedWind.py .                                                                                                                          [ 78%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_Utils.py ...                                                                                                                             [ 87%]
libraries\PyCoxMunk\pycoxmunk\Tests\test_WaterConsts.py ....                                                                                                                      [100%]

---------- coverage: platform win32, python 3.11.2-final-0 -----------
Coverage HTML written to dir htmlcov
```

As you found out, using `numpy.random.uniform` was problematic for the reproducibility. If `numpy.random` is not used with a fixed predefined seed, the unit tests might be working with different data depending on how the default seed is being generated.

The notebooks for GOES and SLSTR now work without any extra work apart from dropping the downloaded files in the correct place inside the `Examples` folder. For the SEVIRI example I keep getting the former error:
```
Cell In[7], line 2
      1 # Load the land mask
----> 2 lsm_scn = Scene([lsm_fname], reader='generic_image')
      3 lsm_scn.load(['image'])

File c:\Users\vic\Desktop\scratch\joss311\Lib\site-packages\satpy\scene.py:133, in Scene.__init__(self, filenames, reader, filter_parameters, reader_kwargs)
    130 if filenames:
    131     filenames = convert_remote_files_to_fsspec(filenames, storage_options)
--> 133 self._readers = self._create_reader_instances(filenames=filenames,
    134                                               reader=reader,
    135                                               reader_kwargs=cleaned_reader_kwargs)
    136 self._datasets = DatasetDict()
    137 self._wishlist = set()

File c:\Users\vic\Desktop\scratch\joss311\Lib\site-packages\satpy\scene.py:154, in Scene._create_reader_instances(self, filenames, reader, reader_kwargs)
    149 def _create_reader_instances(self,
    150                              filenames=None,
    151                              reader=None,
    152                              reader_kwargs=None):
    153     """Find readers and return their instances."""
--> 154     return load_readers(filenames=filenames,
    155                         reader=reader,
    156                         reader_kwargs=reader_kwargs)
...
    207         )
    208     backend = engines[engine]
    209 elif isinstance(engine, type) and issubclass(engine, BackendEntrypoint):

ValueError: unrecognized engine rasterio must be one of: ['netcdf4', 'h5netcdf', 'scipy', 'store', 'zarr']
```

It seems I should be up to date with the dependencies, but anyway I am not an expert on `satpy` and related libraries:
```
(joss311) [vic@onyx] C:\Users\vic\Desktop\scratch> python -c "import sys; [__import__(lib) and print('{0}: {1}'.format(lib, sys.modules[lib].__version__)) for lib in ('satpy', 'rasterio', 'xarray')]"
satpy: 0.42.1
rasterio: 1.3.6
xarray: 2023.4.2
```
