### 2023-05-05 #5074 @simonrp84
Aah, silly me - there's an additional dependency for this, you need to install `rioxarray`!

I'll add that to the note at the top of the SEVIRI example.
