### 2023-05-05 #5074 @molinav
Now the SEVIRI example also works completely for me, it was the missing `rioxarray` as you indicated. There are no more comments from my side. Nice work!
![figure](https://user-images.githubusercontent.com/9979942/236431073-c8f8272f-eae0-49dc-955d-afce9ee731dc.png)
