### 2023-05-26 #5074 @simonrp84
Have now added DOIs to all the references that should have one. Editorialbot seems a bit stuck but hopefully it'll generate the new PDF at some point...
