### 2023-05-26 #5074 @simonrp84
I updated the paper and added DOIs, editorialbot seems happy with it now.
