### 2023-05-26 #5074 @pdebuyl
Thanks. Do you have a recommendation for a geotiff viewer?
