### 2023-05-26 #5074 @simonrp84
I typically use [QGIS ](https://qgis.org/en/site/)for viewing geotiffs.
