### 2023-06-02 simonrp84/PyCoxMunk#5 @pdebuyl
I can't reproduce either (for info)
