### 2023-06-02 #5074 @pdebuyl
Hi Simon, sorry again for the delay.

My comments on the paper:

- Could you say, if appropriate, a few words about how the Fortran -> Python change was done?
- Line 57 starts with "In this case ..." but the sentence does not follow one that could be referred to.
- Can you give a reference for the formula on line 64? If it is common knowledge, a textbook would do for instance, else an article is fine.
- The reference "Sayer 2010" has "(A)ATSR" will a lowercase A and an extra space between "(A)" and "ATSR", can you fix it?
- Nowhere in the article is there a mention of numba. It is used in the code and is the only notable import that is not mentioned.

Regarding the code itself, I could test all example fine now. Thanks for including test data, this makes it more convenient. For info, to read the data I ended up using `rasterio.read` in the notebooks.

A final comment/question/request: In the article, you show a natural color Himawari image and the corresponding reflectance. The algorithm produces a reflectance per channel, what is plotted in Fig 1 (b) ? Is the code to make the figure available? Same question for the image in the docs.
