### 2023-06-06 c9ea4f2
Update the docs to provide more info about the SEVIRI illustrative image on the index page.
