### 2023-06-06 #5074 @simonrp84
> Hi Simon, sorry again for the delay.

No problem, @pdebuyl, thanks for your comments. :)

> My comments on the paper:
>
> * Could you say, if appropriate, a few words about how the Fortran -> Python change was done?

I've edited this section now, as it's not so much a conversion as a newly written code using the fortran as reference to ensure the output is correct.

> * Line 57 starts with "In this case ..." but the sentence does not follow one that could be referred to.

Oops, good catch. Corrected.

> * Can you give a reference for the formula on line 64? If it is common knowledge, a textbook would do for instance, else an article is fine.

The formula is from Sayer 2010, I've added a citation now.

> * The reference "Sayer 2010" has "(A)ATSR" will a lowercase A and an extra space between "(A)" and "ATSR", can you fix it?

Hopefully fixed now.

> * Nowhere in the article is there a mention of numba. It is used in the code and is the only notable import that is not mentioned.

I've added a sentence and corresponding citation that states _pycoxmunk_ uses _numba_ in the BRDF calculations.

>
> Regarding the code itself, I could test all example fine now. Thanks for including test data, this makes it more convenient. For info, to read the data I ended up using `rasterio.read` in the notebooks.

Great! Thanks for info regarding rasterio, I don't have much experience of using it - hence the method in the notebooks - but may change in a future release once I become more familiar with it.

> A final comment/question/request: In the article, you show a natural color Himawari image and the corresponding reflectance. The algorithm produces a reflectance per channel, what is plotted in Fig 1 (b) ? Is the code to make the figure available? Same question for the image in the docs.

The image in the paper is an RGB composite of the Cox-Munk reflectance in the true color channels for Himawari. i.e: The calculated reflectance in the 470, 510 and 640nm channels. I've updated the figure caption to state this.

The image in the docs was produced with an old version of the existing SEVIRI example. The example notebook has been updated and now includes the code needed to generate the image. I've also updated the image itself, as it was produced with an old version of pycoxmunk.
