### 2023-06-06 #5074 @molinav
@simonrp84 Where can I find the Markdown file? I cannot manage to find it in the `joss-papers` or `pycoxmunk` repos, I could try to take a look.
