### 2023-06-06 #5074 @simonrp84
@molinav It's in a [branch of the pycoxmunk repo](https://github.com/simonrp84/PyCoxMunk/tree/paper). Filename is `paper.md`.

Although actually I think this is a bibtex issue with capitalisation, let's see if my new push has fixed it...
