### 2023-06-06 #5074 @molinav
Thanks @simonrp84! I am not sure if it will work in this case, but sometimes I need to enclose acronyms in BibTeX entries with parentheses to ensure that the uppercase/lowercase is preserved as written in the bibfile:

```
@article{sayer:2010,
  title={A sea surface reflectance model for {(A)ATSR}, and application to aerosol retrievals},
  author={Sayer, AM and Thomas, GE and Grainger, RG},
  journal={Atmospheric Measurement Techniques},
  volume={3},
  number={4},
  pages={813--838},
  year={2010},
  publisher={Copernicus GmbH},
  doi = {10.5194/amt-3-813-2010},
}
```
