### 2023-06-08 #5074 @pdebuyl
Thanks @molinav for the help :-)

Thanks @simonrp84 for the update, this is all fine.

Can you make a zenodo archive and post the DOI of the archive here, as well as the specific version of that archive?

The archive must have same title and metadata as the paper, including the author list.
