### 2023-06-09 #5074 @simonrp84
Hi @pdebuyl, I've made a new release and it's automatically generated a zenodo archive:
https://doi.org/10.5281/zenodo.8020079
Version 1.1.0
